requirements.txt
grpcio==1.8.4
six==1.10.0
tensorflow==1.4.1
tensorflow-transform==0.1.10
tensorflow-serving-api==1.4.0
opencv-python==3.3.0.10
opencv-contrib-python==3.3.0.10
scikit-learn==0.19.0
sklearn
statsmodels==0.8.0
pandas==0.20.3
jupyter==1.0.0
grpc-google-iam-v1==0.11.1
google-cloud-storage==1.4.0
google-compute-engine==2.7.0
matplotlib==2.1.0
seaborn==0.8.1
scipy==0.19.1
